requirements.txt
Flask
python-dotenv
openai
torch
scikit-learn
numpy
anthropic
